requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.0.5
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.6.7
decorator==5.1.1
executing==0.8.3
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
grpcio==1.66.0
grpcio-tools==1.66.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.0
jupyter_core==5.7.2
langchain==0.2.14
langchain-core==0.2.34
langchain-huggingface==0.0.3
langchain-ollama==0.1.1
langchain-qdrant==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.104
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
ollama==0.3.1
orjson==3.10.7
packaging==24.1
parso==0.8.3
pillow==10.4.0
pip==24.2
platformdirs==3.10.0
portalocker==2.10.1
prompt-toolkit==3.0.43
protobuf==5.27.3
psutil==5.9.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.15.1
python-dateutil==2.9.0.post0
pywin32==305.1
PyYAML==6.0.2
pyzmq==25.1.2
qdrant-client==1.11.0
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.2.0
sympy==1.13.2
tbb==2021.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing_extensions==4.11.0
urllib3==2.2.2
wcwidth==0.2.5
wheel==0.43.0
Wikipedia-API==0.6.0
yarl==1.9.4
